Remove Erdblick model, view service
