+++ erdblick_app/app/app.component.ts
@@ -6,7 +6,6 @@ import {ActivatedRoute, Params, Router} from "@angular/router";
 import {ParametersService} from "./parameters.service";
 import {OverlayPanel} from "primeng/overlaypanel";
 import {StyleService} from "./style.service";
-import {ViewService} from "./view.service";
 
 @Component({
     selector: 'app-root',
@@ -50,7 +49,6 @@ export class AppComponent {
                 private router: Router,
                 private activatedRoute: ActivatedRoute,
                 public mapService: MapService,
-                public viewService: ViewService,
                 public styleService: StyleService,
                 public jumpToTargetService: JumpTargetService,
                 public parametersService: ParametersService) {
@@ -66,10 +64,10 @@ export class AppComponent {
             this.parametersService.parseAndApplyParams(params, this.firstParamUpdate);
             if (this.firstParamUpdate) {
                 this.firstParamUpdate = false;
-                this.mapService.mapModel.update();
+                this.mapService.update();
                 this.styleService.reapplyAllStyles();
             }
-            setTimeout(() => { this.viewService.viewportToBeUpdated.next(true) }, 1000);
+            setTimeout(() => { this.parametersService.viewportToBeUpdated.next(true) }, 1000);
         });
 
         this.parametersService.parameters.subscribe(parameters => {
